+++ UnitTests/BraintreeVenmoTests/BTVenmoRequest_Tests.swift
@@ -1,15 +1,15 @@
 import XCTest
-import BraintreeVenmo
+@testable import BraintreeVenmo
 
 class BTVenmoRequest_Tests: XCTestCase {
 
     func testPaymentMethodUsageAsString_whenPaymentMethodUsageIsMultiUse_returnsMultiUse() {
         let request = BTVenmoRequest(paymentMethodUsage: .multiUse)
-        XCTAssertEqual(request.paymentMethodUsageAsString, "MULTI_USE")
+        XCTAssertEqual(request.paymentMethodUsage.stringValue, "MULTI_USE")
     }
 
     func testPaymentMethodUsageAsString_whenPaymentMethodUsageIsSingleUse_returnsSingleUse() {
         let request = BTVenmoRequest(paymentMethodUsage: .singleUse)
-        XCTAssertEqual(request.paymentMethodUsageAsString, "SINGLE_USE")
+        XCTAssertEqual(request.paymentMethodUsage.stringValue, "SINGLE_USE")
     }
 }