+++ UnitTests/BraintreeVenmoTests/BraintreeVenmoTests-Bridging-Header.h
@@ -1,3 +1,0 @@
-#import <BraintreeVenmo/BTVenmoClient_Internal.h>
-#import <BraintreeCore/BraintreeCore-Swift.h>
-#import <BraintreeVenmo/BraintreeVenmo-Swift.h>